identified some issues
namely that the listification inflates the log sizes (any update to a
block doesn't just edit that key, it rewrites the entire block, due to
the hashed structure. Maybe we find a better way based off of inspecting
UUIDs of list items, when available?) anyway this causes some major
rewrites of that structure which surface some issues.